docs: Better documentation of trusting SSL certs (#215)
DH-18191: Better documentation of trusting SSL certs

NodeJS doesn't use the OS cert store to determine which signing
certificates it trusts. Instead, it has a static list of trusted certs
included in the install. For certificates not in this list, there is a
`NODE_EXTRA_CA_CERTS` environment variable that can be set to tell
NodeJS about additional certs to trust. This PR updates the docs with
instructions of how to set this.

> Note there is some indication that extensions may eventually allow
using Chrome's `fetch`. If / when that happens, it might be possible to
use that for Deephaven extension instead of our current grpc transport
implementation that uses `node:http2` module. This could allow
extensions to use the OS cert store instead of this env variable.
https://github.com/microsoft/vscode/issues/12588#issuecomment-2111861237