fix: DH-19010: http2 errors (#228)
DH-19010:
* Handle client disconnects and proper disposal of DhService (Note this
depends on changes made in
https://github.com/deephaven/web-client-ui/pull/2400 which have been
brought in by @deephaven/jsapi-nodejs@0.107.0)
* Added support for `DHC_PACKAGES_PATH` env variable. If set, this
aliases `@deephaven/*` packages to the specified path
* Upgraded dhc loader to use the simpler `loadDhModules` function

### Testing
#### Disconnect Handling
To test the disconnect handling, you can introduce this code block after
the `new dhc.CoreClient` block in `ExtensionController.ts`
```typescript
// Destroy the http2 session after 2 seconds to mimic disconnect outside
// of jsapi
setTimeout(() => {
  for (var session of (
    NodeHttp2gRPCTransport as any
  ).sessionMap.values()) {
    session.destroy();
  }
}, 2000);
```

1. Click on Community server node to connect
2. The added codeblock should destroy the http2 session after 2 seconds
which will cause client to disconnect and the CONNECTION node to be
removed
3. Should be able to reconnect (it will disconnect 2 seconds later, but
we are proving that a disconnect can recover)

#### DHC_PACKAGES_PATH
The `DHC_PACKAGES_PATH` env variable can be tested by creating a
`.env.local` file and setting the it to the path of a local DHC web
`/packages` directory.

> Note that if you have previously run the debugger, you will need to
kill the terminals for the build tasks in order to get a fresh build.