docs: DOC-697 - Update VS Code extension README (#233)
This adds a hook/more info about DH to the top of the README, as well as
rework the guide to improve clarity.

Adding Chip as a reviewer, but only because he was part of discussions
regarding readability. He should not need to actually review this unless
we can't agree on something.