fix: DH-20899: Tree view command errors (#270)
DH-20899: Fixed some treeview command errors

- The refresh button in title bar of PANELS panel was completely broken
due to using incorrect command. I created a dedicated command for this.
- Inline actions for treeview items sometimes pass undefined instead of
the treeview item. I've mostly observed this when changing to another
window and coming back to VS Code. It's possible there's some treeview
data sync issue like the tree in the middle of a refresh or something,
but it seems like a VS Code bug. Regardless, the fix was to ignore
undefined values passed to command handlers and user can just click
again instead of getting an error toast which was happening before
- I also fixed a minor issue where package.json showed validation
warnings due to missing icons. These get used if a user drags a panel
out of the DH view into the activity bar as its own panel. For now just
re-using the DH icon. Could always have dedicated icons for each panel,
but this seems very low priority and out of scope.