fix: DH-21138: Use gRPC transport with DHE client (#274)
DH-21138:
- Use gRPC transport with DHE client
- I also slipped in a 1 line fix to the default value being used by the
query creation UI. It stores the last selected language, but prefers the
language of the active editor. In cases where no files are open, it was
defaulting to undefined instead of using the last selection. This is
fixed now

## Testing
I tested connection scenarios:
For
- grizzly envoy
- grizzly non-envoy
- gplus envoy
- gplus non-envoy

I tested basic login, keypair gen and keypair login

Verified that the last language selected is remembered when no files are
open and that if file is open, it's language is used as default for
query creation dropdown